Modify the 'canonical' moves to map to the right face.
If the puzzle gets reoriented, the GAN faces will no longer line
up with the model faces. Take the 'canonical' moves coming from the
GAN cube (which assume a White up, Green front orientation) and map
them to the correct faces in the current state, so that in the
future when this code also emits M, E, S and x, y, z moves the
correct faces will twist even though the centers are reoriented.