First pass at orientation tracking.
Log debug output that corresponds to correctly computed cube
rotations. The approach is to create a quaternion for each facing
direction, and compare the current orientation of the cube to that
direction to discover whether the orientation is close enough to
that facing to say the user has rotated the cube. Use a small dead
zone around each facing so that the rotation is stable, so the user
has to go slightly past 45 degrees towards the new face to trigger
the change.